+++ src/endpoints/chats.js
@@ -9,7 +9,14 @@ import { sync as writeFileAtomicSync } from 'write-file-atomic';
 import _ from 'lodash';
 
 import { jsonParser, urlencodedParser } from '../express-common.js';
-import { getConfigValue, humanizedISO8601DateTime, tryParse, generateTimestamp, removeOldBackups } from '../util.js';
+import {
+    getConfigValue,
+    humanizedISO8601DateTime,
+    tryParse,
+    generateTimestamp,
+    removeOldBackups,
+    formatBytes,
+} from '../util.js';
 
 const isBackupDisabled = getConfigValue('disableChatBackup', false);
 const maxTotalChatBackups = Number(getConfigValue('maxTotalChatBackups', -1));
@@ -46,6 +53,9 @@ function backupChat(directory, name, chat) {
     }
 }
 
+/**
+ * @type {Map<string, import('lodash').DebouncedFunc<function(string, string, string): void>>}
+ */
 const backupFunctions = new Map();
 
 /**
@@ -57,20 +67,7 @@ function getBackupFunction(handle) {
     if (!backupFunctions.has(handle)) {
         backupFunctions.set(handle, _.throttle(backupChat, throttleInterval, { leading: true, trailing: true }));
     }
-    return backupFunctions.get(handle);
-}
-
-/**
- * Formats a byte size into a human-readable string with units
- * @param {number} bytes - The size in bytes to format
- * @returns {string} The formatted string (e.g., "1.5 MB")
- */
-function formatBytes(bytes) {
-    if (bytes === 0) return '0 B';
-    const k = 1024;
-    const sizes = ['B', 'KB', 'MB', 'GB'];
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
+    return backupFunctions.get(handle) || (() => {});
 }
 
 /**
@@ -710,7 +707,7 @@ router.post('/search', jsonParser, function (request, response) {
         }
 
         // Sort by last message date descending
-        results.sort((a, b) => new Date(b.last_mes) - new Date(a.last_mes));
+        results.sort((a, b) => new Date(b.last_mes).getTime() - new Date(a.last_mes).getTime());
         return response.send(results);
 
     } catch (error) {
